move log/logger.go to third_party and log/test/logger.go to test/log/logger.go
