+++ da/da.go
@@ -6,7 +6,7 @@ import (
 
 	ds "github.com/ipfs/go-datastore"
 
-	"github.com/rollkit/rollkit/log"
+	"github.com/rollkit/rollkit/third_party/log"
 	"github.com/rollkit/rollkit/types"
 )
