+++ da/test/da_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/rollkit/rollkit/da/grpc/mockserv"
 	"github.com/rollkit/rollkit/da/mock"
 	"github.com/rollkit/rollkit/da/registry"
-	"github.com/rollkit/rollkit/log/test"
 	"github.com/rollkit/rollkit/store"
+	test "github.com/rollkit/rollkit/test/log"
 	"github.com/rollkit/rollkit/types"
 )
