+++ p2p/client.go
@@ -28,7 +28,7 @@ import (
 	tmcrypto "github.com/tendermint/tendermint/crypto"
 
 	"github.com/rollkit/rollkit/config"
-	"github.com/rollkit/rollkit/log"
+	"github.com/rollkit/rollkit/third_party/log"
 )
 
 // TODO(tzdybal): refactor to configuration parameters
